Add promos siblings, add order cancel, update price calculation and presenting, add twin helpers

+++ Sources/App/Migrations/CreateMakeupers.swift
@@ -13,7 +13,7 @@ struct CreateMakeupers: Migration {
             .id()
             .field("name", .string)
             .field("platform_ids", .array(of: .json))
-            .field("price", .int, .required)
+            .field("price", .float, .required)
             .create()
     }
