+++ Sources/App/Types/Modeled/Studio.swift
@@ -27,7 +27,7 @@ final class Studio: StudioProtocol {
 
     var photos: [PlatformFileModel] = []
 
-    var price: Int = 0
+    var price: Float = 0
     
     required init() {}
